rename rules::action to rules::action_only
